# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: HW2 for Web Api
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /deletes:
    delete:
      description: Server accepts only delete requests
      # used as the method name of the controller
      operationId: deletes
      parameters:
        - name: name
          in: query
          description: The delete request to a server
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Response"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Response:
    required:
      - message
    properties:
      message:
        type: string
  Error:
    required:
      - message
    properties:
      message:
        type: string